requirements.txt
pandas
Numpy
requests
seaborn
psycopg2-binary
python-dotenv
openpyxl
xlrd
sqlalchemy